image_types_qcom: generate flashable images for iq-x7181-evk
Spinor firmware is needed on iq-x7181 based devices, use the external
released non-hlos images and deploy them to generate the flat build for
spinor flash. Meanwhile, copy the programer outside of spinor folder to
be part of NVMe flat build.

+++ classes-recipe/image_types_qcom.bbclass
@@ -19,6 +19,7 @@ QCOM_DTB_FILE ?= "dtb.bin"
 
 QCOM_BOOT_FILES_SUBDIR ?= ""
 QCOM_PARTITION_FILES_SUBDIR ??= "${QCOM_BOOT_FILES_SUBDIR}"
+QCOM_PARTITION_FILES_SUBDIR_SPINOR ??= ""
 
 QCOM_PARTITION_CONF ?= "qcom-partition-conf"
 
@@ -89,7 +90,8 @@ create_qcomflash_pkg() {
     fi
 
     if [ -n "${QCOM_BOOT_FILES_SUBDIR}" ]; then
-        if [ -n "${QCOM_CDT_FILE}" ]; then
+        # check CDT file existence and install as per need, for targets with spinor, CDT file will be in spinor subfolder instead of root folder.
+        if [ -n "${QCOM_CDT_FILE}" ] && [ -e "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${QCOM_CDT_FILE}.bin" ]; then
             install -m 0644 ${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${QCOM_CDT_FILE}.bin cdt.bin
         fi
 
@@ -123,6 +125,36 @@ create_qcomflash_pkg() {
     # abl2esp
     if [ -e "${DEPLOY_DIR_IMAGE}/abl2esp-${ABL_SIGNATURE_VERSION}.elf" ]; then
         install -m 0644 "${DEPLOY_DIR_IMAGE}/abl2esp-${ABL_SIGNATURE_VERSION}.elf" .
+    fi
+
+    # spinor firmware, partition bins, CDT etc.
+    if [ -d "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" ]; then
+        install -d spinor
+        find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" -maxdepth 1 -type f -exec install -m 0644 {} spinor \;
+
+        # partition bins for spinor
+        for pbin in `find ${DEPLOY_DIR_IMAGE}/${QCOM_PARTITION_FILES_SUBDIR_SPINOR} -maxdepth 1 -type f -name 'gpt_main*.bin' \
+                    -o -name 'gpt_backup*.bin' -o -name 'patch*.xml'`; do
+            install -m 0644 ${pbin} spinor
+        done
+
+        # skip BLANK_GPT and WIPE_PARTITIONS for rawprogram xml files
+        for rawpg in `find ${DEPLOY_DIR_IMAGE}/${QCOM_PARTITION_FILES_SUBDIR_SPINOR} -maxdepth 1 -type f -name 'rawprogram*.xml' \
+                    ! -name 'rawprogram*_BLANK_GPT.xml' ! -name 'rawprogram*_WIPE_PARTITIONS.xml'`; do
+            install -m 0644 ${rawpg} spinor
+        done
+
+        for zeros in `find ${DEPLOY_DIR_IMAGE}/${QCOM_PARTITION_FILES_SUBDIR_SPINOR} -maxdepth 1 -type f -name 'zeros_*.bin'`; do
+            install -m 0644 ${zeros} spinor
+        done
+
+        # rename CDT file to match the filename in partition conf
+        if [ -n "${QCOM_CDT_FILE}" ]; then
+            install -m 0644 ${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor/cdt.bin spinor/cdt.bin
+        fi
+
+        # copy programer to support flash of hlos images
+        find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" -maxdepth 1 -type f -name 'xbl_s_devprg_ns.melf' -exec install -m 0644 {} . \;
     fi
 
     # Create symlink to ${QCOMFLASH_DIR} dir
